Refactoring the code in monai.transforms.lazy.executors into monai.transforms.lazy.functional
