+++ monai/transforms/compose.py
@@ -27,7 +27,7 @@ from monai.config import NdarrayOrTensor
 from monai.transforms.inverse import InvertibleTransform
 
 # For backwards compatibility (so this still works: from monai.transforms.compose import MapTransform)
-from monai.transforms.lazy.executors import apply_pending_transforms
+from monai.transforms.lazy.functional import apply_pending_transforms
 from monai.transforms.traits import ThreadUnsafe
 from monai.transforms.transform import (  # noqa: F401
     LazyTransform,
