+++ monai/transforms/lazy/utils.py
@@ -11,11 +11,12 @@
 
 from __future__ import annotations
 
+from typing import Any, Sequence, Mapping
+
 import warnings
 
 import numpy as np
 import torch
-
 import monai
 from monai.config import NdarrayOrTensor
 from monai.data.utils import AFFINE_TOL
